backend/pyproject.toml
[project]
name = "edubot-backend"
version = "1.0.0"
description = "Multi-Model AI Chatbot with LangGraph Agent Workflow"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core Framework
    "fastapi==0.115.5",
    "uvicorn[standard]==0.34.0",
    "pydantic[all,email]==2.12.0",
    "pydantic-settings==2.11.0",
    
    # Database
    "sqlalchemy==2.0.43",
    "psycopg2-binary==2.9.11",
    "asyncpg==0.30.0",
    "alembic==1.17.0",
    
    # Authentication
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "bcrypt<4.1",
    
    # Environment & Config
    "python-dotenv==1.1.1",
    
    # AI/ML Core - LangChain & LangGraph
    "langchain==0.3.27",
    "langchain-core==0.3.78",
    "langchain-openai==0.3.0",
    "langchain-google-genai==2.1.12",
    "langchain-ollama==0.3.10",
    "langchain-community==0.3.31",
    
    # LangGraph (Agent Workflow & State Management)
    "langgraph==0.6.8",
    "langgraph-checkpoint==2.1.2",
    "langgraph-checkpoint-postgres==2.0.10",
    
    # OpenAI
    "openai==1.58.1",
    
    # Google AI
    "google-generativeai==0.8.3",
    "google-api-core==2.25.2",
    "google-auth==2.41.1",
    
    # Utilities
    "httpx==0.28.1",
    "requests==2.32.5",
    "tenacity==9.1.2",
    "python-multipart==0.0.20",
    
    # CORS
    "python-cors==1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
